requirements.txt
openai
langchain
langchain-openai
langchain-experimental
python-dotenv
PyPDF2
pandas
streamlit
tiktoken
